initial draft
A re-write of an old script I wrote ‎June ‎12th ‎2021, this new maxscript focuses just on the v29 of DTS (Dynamix Three Space) file format found in BoneTown, which is called (.dwc). It's unclear if the DTS is modified or original to the Torque3D Game Engine's original DTS format specifications since the (TGE) website only provides a basic format specification for v26.

Anyway, imports meshes from dwc files, materials look a bit off but I decided to stop work on this project when I later realized exporting models back would be problematic. more specifics about that are written in the notes section in the actual maxscript.